26 Image resizing logic
Use slider to scale the width of the image from the size of the container as a minimum as the size of the image
as a maximum.

Include a stub photo in the project that is hard coded into the component.

Add styles to create the effect that the image is being resized.