FK-26 Upload images asynchronously
Add Paperclip and Mongoid Paperclip for saving images and to a mongo database.

Use form data in sync method of events to send back a multipart/form-data request including the
event image now.

NOTE: to use paperclip it was necessary to install ImageMagick. This was installed by
sudo apt-get install imagemagick.

Added paperclip's file directory to the repository ignore list for versioning.