FK-26 Link up url for thumbnail image
Use the url generated by paperclip to display a thumbnail image in the event list.

Change image settings of paperclip to not mantain image resolution, just take a chunk out
of the image instead to match the required size.

Use the old default image as the missing image for paperclip instead of a default.