Issue #31 Create the event page and give a trigger
Add an event trigger on the event name in the collapsed event to open a single event.
Add an event in the event list module to signal the app to open the event page.

Add an event page module and have it listen for the event show event.
Create a simple template for an event shown full page.