Issue #31 Refactor: back the image trimmer w model
Introduce a model that can be used to track the state of the image in the image
trimmer.
Save changes to the image to a model instead of through the controller.
This is part of a longer refactor to depend on a model and be able
to recreate the state of the image trimmer.