Issue #22 Add logic to add events to a block
Introduce the add events method to the event block.
Introduce the event limit on the event block.
Use the event limit to taper the number of events that can be added to a block.
Set default event limit to 3.
Introduce increase limit method to bump up the number of events that can be on a block.