Issue #129 Made the server just return events
Instead of trying to determine exactly how many of the events to return
and only returning those, we now return all the events in the first 3
days on page load, and then all the events with in a range upon each
subsequent fetch on scroll.

We have also removed a refresh so the back button does not reset the list
of events.