[#73023496] Refactor subkasts to use models
Remove the subkast definition hashes in kiwi.
Introduce models for subkasts.
Replace all references to subkasts and have them use the new subkasts
collection.
Populate the subkasts collection with all subkasts provided by the
server side subkasts models.
Refactor some of the app to not use requests inside of modules, but use
the config passed to them on init instead.